bot: Add mentions param for /quote list
