+++ Bot/Commands/SearchCommands.cs
@@ -61,7 +61,8 @@ namespace Bot.Commands
         {
             if (args.Author.IsBot) return;
 
-            var commands = _commands.Keys.Where(commName => args.Message.Content.ToUpper().Contains(commName)).ToArray();
+            var commands = _commands.Keys.Where(commName => args.Message.Content.ToUpper().Contains(commName))
+                .ToArray();
 
             if (!commands.Any())
             {
@@ -81,36 +82,39 @@ namespace Bot.Commands
                 catch (Exception e)
                 {
                     Logger.LogError(e, "Failed to execute command {Command}", commName);
-                    await args.Message.RespondAsync("I was supposed to send you a message, but I failed somewhere. Please try again or contact the developer.");
+                    await args.Message.RespondAsync(
+                        "I was supposed to send you a message, but I failed somewhere. Please try again or contact the developer.");
                 }
             }
         }
-    }
 
-    public class SansCarSearchCommand : ISearchCommand
-    {
-        private ILogger<BaseDiscordClient> _logger;
-        public string GetName()
-        {
-            return "Sans car";
-        }
 
-        public string[] GetAliases()
+        public class SansCarSearchCommand : ISearchCommand
         {
-            return new[] { "sans car", "sanscar" };
-        }
+            private ILogger<BaseDiscordClient> _logger;
 
-        public void SetLogger(ILogger<BaseDiscordClient> logger)
-        {
-            _logger = logger;
-        }
+            public string GetName()
+            {
+                return "Sans car";
+            }
 
-        public async Task Execute(DiscordClient sender, MessageCreateEventArgs args)
-        {
-            await using var imageFile = MediaResources.GetImage("sans-car.jpg");
-            var message = new DiscordMessageBuilder()
-                .WithFile(imageFile);
-            await args.Message.RespondAsync(message);
+            public string[] GetAliases()
+            {
+                return new[] { "sans car", "sanscar" };
+            }
+
+            public void SetLogger(ILogger<BaseDiscordClient> logger)
+            {
+                _logger = logger;
+            }
+
+            public async Task Execute(DiscordClient sender, MessageCreateEventArgs args)
+            {
+                await using var imageFile = MediaResources.GetImage("sans-car.jpg");
+                var message = new DiscordMessageBuilder()
+                    .WithFile(imageFile);
+                await args.Message.RespondAsync(message);
+            }
         }
     }
